Added more procedures for conditioning data and wrote an sklearn-based decision tree classifier
The decision tree classifier makes a binary classification (rainfall == 0 or
rainfall > 0), with the idea that a second learner will be applied to predict
the cdf for examples that are classisfied into the rainfall > 0 class.

As preprocessing, all time series are Riemann summed. The number of time
points per example has been incorporated as an additional attribute, because it
appears to correlate with expected rainfall amount.